quick and dirty LED layout visualization
i'm planning on writing a snake-game that runs directly on the kbd,
and I wanted a way to get a feel for how the leds are layed out.
It's not a perfect grid, so there is some cell connection topology
to experiment with.

I'm probably going to base the code off of the LED effects under
`reywood`, since he seems to have a good way of addressing LEDs